[refactor] 진행중인 모임방 response 필드 타입 수정

+++ src/main/java/konkuk/thip/room/adapter/in/web/response/RoomPlayingDetailViewResponse.java
@@ -22,7 +22,7 @@ public record RoomPlayingDetailViewResponse(
         String bookTitle,
         String authorName,
         int currentPage,
-        double userPercentage,
+        int userPercentage,
         List<CurrentVote> currentVotes
 ) {
     public record CurrentVote(
